+++ Tools/LambdaTestTool-v2/src/Amazon.Lambda.TestTool/Models/Exceptions.cs
@@ -18,3 +18,11 @@ public abstract class TestToolException(string message, Exception? innerExceptio
 /// <param name="innerException">The inner exception, if any.</param>
 public class InvalidApiGatewayModeException(string message, Exception? innerException = null)
     : TestToolException(message, innerException);
+
+/// <summary>
+/// Thrown if the save request is invalid.
+/// </summary>
+/// <param name="message">The message used in the exception.</param>
+/// <param name="innerException">The inner exception, if any.</param>
+public class InvalidSaveRequestException(string message, Exception? innerException = null)
+    : TestToolException(message, innerException);