+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Release 2025-06-25
+
+### Amazon.Lambda.TestTool (0.11.0)
+* Add the ability to save requests and hide UI elements
+
 ## Release 2025-06-03
 
 ### Amazon.Lambda.RuntimeSupport (1.13.1)